End screen, Some bug fixes, Lion Flame Trail.
I fixed some bugs from pausing the screen, I was able to put a isPaused
Boolean and now you cannot buy upgrades from the pause screen. I have
when you die you go to the game over screen with highscores, we need to
tighten it up but it looks good. I put a flame trail on the Lion.